+++ large_image/tilesource/base.py
@@ -2411,7 +2411,10 @@ class TileSource:
                 pixel['tile'] = tile
             img = tile['tile']
             if img.size[0] >= 1 and img.size[1] >= 1:
-                pixel.update(dict(zip(img.mode.lower(), img.load()[0, 0])))
+                if len(img.mode) > 1:
+                    pixel.update(dict(zip(img.mode.lower(), img.load()[0, 0])))
+                else:
+                    pixel.update(dict(zip([img.mode.lower()], [img.load()[0, 0]])))
         return pixel